Add files via upload
This info system is based object-oriented programming. It has three main parts: customer, medicine stock and supplier. For each in case if the info system is used the first time it automatically creates text document ( the user enters specific data in it). The system has smart menu, editing and search options.